move param parsing to option_utils and remove unused functions
