+++ gwin/option_utils.py
@@ -19,6 +19,7 @@
 
 import logging
 import shutil
+import argparse
 
 from pycbc import (conversions, inject, transforms)
 from pycbc.distributions import (bounded, constraints)
@@ -30,9 +31,6 @@ from pycbc.strain import from_cli_multi_ifos as strain_from_cli_multi_ifos
 from pycbc.strain import (gates_from_cli, psd_gates_from_cli,
                           apply_gates_to_td, apply_gates_to_fd)
 from pycbc import waveform
-from gwin import (burn_in, models, sampler)
-from gwin.io.hdf import InferenceFile, check_integrity
-from gwin.io.txt import InferenceTXTFile
 
 
 # -----------------------------------------------------------------------------
@@ -65,124 +63,6 @@ def config_parser_from_cli(opts):
     else:
         overrides = None
     return WorkflowConfigParser(opts.config_files, overrides)
-
-
-# -----------------------------------------------------------------------------
-#
-#                    Utilities for setting up a sampler
-#
-# -----------------------------------------------------------------------------
-
-def add_sampler_option_group(parser):
-    """Adds the options needed to set up an inference sampler.
-
-    Parameters
-    ----------
-    parser : object
-        ArgumentParser instance.
-    """
-    sampler_group = parser.add_argument_group(
-        "Arguments for setting up a sampler")
-
-    # required options
-    sampler_group.add_argument(
-        "--sampler", required=True, choices=sampler.samplers.keys(),
-        help="Sampler class to use for finding posterior.")
-    sampler_group.add_argument(
-        "--niterations", type=int,
-        help="Number of iterations to perform. If 'use_sampler' is given to "
-             "burn-in-function, this will be counted after the sampler's burn "
-             "function has run. Otherwise, this is the total number of "
-             "iterations, including any burn in.")
-    sampler_group.add_argument(
-        "--n-independent-samples", type=int,
-        help="Run the sampler until the specified number of "
-             "independent samples is obtained, at minimum. Requires "
-             "checkpoint-interval. At each checkpoint the burn-in iteration "
-             "and ACL is updated. The number of independent samples is the "
-             "number of samples across all walkers starting at the "
-             "burn-in-iteration and skipping every `ACL`th iteration. "
-             "Either this or niteration should be specified (but not both).")
-    # sampler-specific options
-    sampler_group.add_argument(
-        "--nwalkers", type=int, default=None,
-        help="Number of walkers to use in sampler. Required for MCMC "
-             "samplers.")
-    sampler_group.add_argument(
-        "--ntemps", type=int, default=None,
-        help="Number of temperatures to use in sampler. Required for parallel "
-             "tempered MCMC samplers.")
-    sampler_group.add_argument(
-        "--burn-in-function", default=None, nargs='+',
-        choices=burn_in.burn_in_functions.keys(),
-        help="Use the given function to determine when chains are burned in. "
-             "If none provided, no burn in will be estimated. "
-             "If multiple functions are provided, will use the maximum "
-             "iteration from all functions.")
-    sampler_group.add_argument(
-        "--min-burn-in", type=int, default=0,
-        help="Force the burn-in to be at least the given number of "
-             "iterations.")
-    sampler_group.add_argument(
-        "--update-interval", type=int, default=None,
-        help="If using kombine, specify the number of steps to take between "
-             "proposal updates. Note: for purposes of updating, kombine "
-             "counts iterations since the last checkpoint. This interval "
-             "should therefore be less than the checkpoint interval, else "
-             "no updates will occur. To ensure that updates happen at equal "
-             "intervals, make checkpoint-interval a multiple of "
-             "update-interval.")
-    sampler_group.add_argument(
-        "--nprocesses", type=int, default=None,
-        help="Number of processes to use. If not given then use maximum.")
-    sampler_group.add_argument(
-        "--use-mpi", action='store_true', default=False,
-        help="Use MPI to parallelize the sampler")
-    sampler_group.add_argument(
-        "--logpost-function", default="logposterior",
-        help="Which attribute of the model to use for the logposterior. "
-             "The default is logposterior. For example, if using the "
-             "gaussian_noise model, you may wish to set this to logplr, since "
-             "the logposterior includes a large constant contribution from "
-             "log noise likelihood.")
-
-    return sampler_group
-
-
-def sampler_from_cli(opts, model, pool=None):
-    """Parses the given command-line options to set up a sampler.
-
-    Parameters
-    ----------
-    opts : object
-        ArgumentParser options.
-    model : model
-        The model to use with the sampler.
-
-    Returns
-    -------
-    gwin.sampler
-        A sampler initialized based on the given arguments.
-    """
-    # create a wrapper for the model
-    model = models.CallModel(model, opts.logpost_function)
-
-    # Used to help paralleize over multiple cores / MPI
-    if opts.nprocesses > 1:
-        models._global_instance = model
-        model_call = models._call_global_model
-    else:
-        model_call = None
-
-    sclass = sampler.samplers[opts.sampler]
-
-    pool = choose_pool(mpi=opts.use_mpi, processes=opts.nprocesses)
-
-    if pool is not None:
-        pool.count = opts.nprocesses
-
-    return sclass.from_cli(opts, model,
-                           pool=pool, model_call=model_call)
 
 
 # -----------------------------------------------------------------------------
@@ -302,6 +182,110 @@ def data_from_cli(opts):
 #                Utilities for plotting results
 #
 # -----------------------------------------------------------------------------
+
+
+class ParseLabelArg(argparse.Action):
+    """Argparse action that will parse arguments that can accept labels.
+
+    This assumes that the values set on the command line for its assigned
+    argument are strings formatted like ``PARAM[:LABEL]``. When the arguments
+    are parsed, the ``LABEL`` bit is stripped off and added to a dictionary
+    mapping ``PARAM -> LABEL``. This dictionary is stored to the parsed
+    namespace called ``{dest}_labels``, where ``{dest}`` is the argument's
+    ``dest`` setting (by default, this is the same as the option string).
+    Likewise, the argument's ``dest`` in the parsed namespace is updated so
+    that it is just ``PARAM``.
+
+    If no ``LABEL`` is provided, then ``PARAM`` will be used for ``LABEL``.
+
+    This action can work on arguments that have ``nargs != 0`` and ``type`` set
+    to ``str``.
+    """
+    def __init__(self, type=str, nargs=None, **kwargs):
+        # check that type is string
+        if type != str:
+            raise ValueError("the type for this action must be a string")
+        if nargs == 0:
+            raise ValueError("nargs must not be 0 for this action")
+        super(ParseLabelArg, self).__init__(type=type, nargs=nargs,
+                                                 **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        singlearg = isinstance(values, (str, unicode))
+        if singlearg:
+            values = [values]
+        params = []
+        labels = {}
+        for param in values:
+            psplit = param.split(':')
+            if len(psplit) == 2:
+                param, label = psplit
+            else:
+                label = param
+            labels[param] = label
+            params.append(param)
+        # update the namespace
+        if singlearg:
+            params = params[0]
+        setattr(namespace, self.dest, params)
+        setattr(namespace, '{}_labels'.format(self.dest), labels)
+
+
+class ParseParametersArg(ParseLabelArg):
+    """Argparse action that will parse parameters and labels from an opton.
+
+    Does the same as ``ParseLabelArg``, with the additional functionality that
+    if ``LABEL`` is a known parameter in ``pycbc.waveform.parameters``, then
+    the label attribute there will be used in the labels dictionary.
+    Otherwise, ``LABEL`` will be used.
+
+    Examples
+    --------
+    Create a parser and add two arguments that use this action (note that the
+    first argument accepts multiple inputs while the second only accepts a
+    single input):
+
+    >>> import argparse
+    >>> parser = argparse.ArgumentParser()
+    >>> parser.add_argument('--parameters', type=str, nargs="+",
+                            action=ParseParametersArg)
+    >>> parser.add_argument('--z-arg', type=str, action=ParseParametersArg)
+
+    Parse a command line that uses these options:
+
+    >>> import shlex
+    >>> cli = "--parameters 'mass1+mass2:mtotal' ra ni --z-arg foo:bar"
+    >>> opts = parser.parse_args(shlex.split(cli))
+    >>> opts.parameters
+    ['mass1+mass2', 'ra', 'ni']
+    >>> opts.parameters_labels
+    {'mass1+mass2': '$M~(\\mathrm{M}_\\odot)$', 'ni': 'ni', 'ra': '$\\alpha$'}
+    >>> opts.z_arg
+    'foo'
+    >>> opts.z_arg_labels
+    {'foo': 'bar'}
+
+    In the above, the first argument to ``--parameters`` was ``mtotal``. Since
+    this is a recognized parameter in ``pycbc.waveform.parameters``, the label
+    dictionary contains the latex string associated with the ``mtotal``
+    parameter. A label was not provided for the second argument, and so ``ra``
+    was used. Since ``ra`` is also a recognized parameter, its associated latex
+    string was used in the labels dictionary. Since ``ni`` and ``bar`` (the
+    label for ``z-arg``) are not recognized parameters, they were just used
+    as-is in the labels dictionaries.
+    """
+    def __call__(self, parser, namespace, values, option_string=None):
+        super(ParseParametersArg, self).__call__(parser, namespace, values,
+                                                 option_string=option_string)
+        # try to replace the labels with a label from waveform.parameters
+        labels = getattr(namespace, '{}_labels'.format(self.dest))
+        for param, label in labels.items():
+            try:
+                label = getattr(waveform.parameters, label).label
+                labels[param] = label
+            except AttributeError:
+                pass
+
 
 def add_plot_posterior_option_group(parser):
     """Adds the options needed to configure plots of posterior results.
@@ -443,7 +427,7 @@ def add_scatter_option_group(parser):
                                               "scatter plot.")
 
     scatter_group.add_argument(
-        '--z-arg', type=str, default=None,
+        '--z-arg', type=str, default=None, action=ParseParametersArg,
         help='What to color the scatter points by. Syntax is the same as the '
              'parameters option.')
     scatter_group.add_argument(
